refactor: rename model_explainability to ai_safety
Rename test suite from model_explainability to ai_safety to better
reflect the broader scope of AI safety testing capabilities.

Changes:
- Rename tests/model_explainability/ -> tests/ai_safety/
- Update all Python imports (tests.model_explainability -> tests.ai_safety)
- Update pytest marker (model_explainability -> ai_safety)
- Update config files: pytest.ini, conftest.py
- Update documentation: AGENTS.md, CONSTITUTION.md, PARALLEL_TESTING.md
- Update CI config: .github/labeler.yml
- Update all pytest markers in test files
- Update file path references in test data constants

Backport to 3.4ea1
