+++ CONSTITUTION.md
@@ -95,7 +95,7 @@ All code MUST consider security implications.
 
 - All tests MUST apply relevant markers from pytest.ini
 - Use tier markers (smoke, sanity, tier1, tier2) to indicate test priority
-- Use component markers (model_explainability, llama_stack, rag) for ownership
+- Use component markers (ai_safety, llama_stack, rag) for ownership
 - Use infrastructure markers (gpu, parallel, slow) for execution filtering
 
 ### Test Organization
