+++ pytest.ini
@@ -49,7 +49,7 @@ markers =
     # Component markers: used to document test suite ownership and to facilitate their execution
     # Example: "pytest -m rag" will run all tests owned by the RAG team, which might include
     # multiple subfolders at /tests/llama_stack
-    model_explainability
+    ai_safety
     llama_stack
     rag
