+++ tests/ai_safety/evalhub/utils.py
@@ -1,0 +1,47 @@
+import requests
+from simple_logger.logger import get_logger
+
+from tests.ai_safety.evalhub.constants import (
+    EVALHUB_HEALTH_PATH,
+    EVALHUB_HEALTH_STATUS_HEALTHY,
+)
+from utilities.guardrails import get_auth_headers
+
+LOGGER = get_logger(name=__name__)
+
+
+def validate_evalhub_health(
+    host: str,
+    token: str,
+    ca_bundle_file: str,
+) -> None:
+    """Validate that the EvalHub service health endpoint returns healthy status.
+
+    Args:
+        host: Route host for the EvalHub service.
+        token: Bearer token for authentication.
+        ca_bundle_file: Path to CA bundle for TLS verification.
+
+    Raises:
+        AssertionError: If the health check fails.
+        requests.HTTPError: If the request fails.
+    """
+    url = f"https://{host}{EVALHUB_HEALTH_PATH}"
+    LOGGER.info(f"Checking EvalHub health at {url}")
+
+    response = requests.get(
+        url=url,
+        headers=get_auth_headers(token=token),
+        verify=ca_bundle_file,
+        timeout=10,
+    )
+    response.raise_for_status()
+
+    data = response.json()
+    LOGGER.info(f"EvalHub health response: {data}")
+
+    assert "status" in data, "Health response missing 'status' field"
+    assert data["status"] == EVALHUB_HEALTH_STATUS_HEALTHY, (
+        f"Expected status '{EVALHUB_HEALTH_STATUS_HEALTHY}', got '{data['status']}'"
+    )
+    assert "timestamp" in data, "Health response missing 'timestamp' field"
